drivers: net: ipq6018: Add optional active_port env support
This patch adds support for active_port env. With this
change in ipq6018, the end user has an option to
enable & maintain only 1 active ethernet port. There
are no multi-port use case in U-Boot and when multiple
ports were enabled for Kernel usecases, there were some
timeout issues in some special cases during crashdump
collection.

active_port can be used to configure the port which
will be used for TFTP download/upload in u-boot in
those setups. Example: To set Port2 as active_port
following command can be used:

setenv active_port 2 && savee

Note that, the active_port range is from 0 to 4 (there
are a total of 5 ports in ipq6018).

Change-Id: I5e41269ebab8eef96efeeb65f1324cacc9d56710
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>