drivers: net: ipq6018: Removing the additional rx buffer allocation
During long run tests on image download in u-boot, 
EDMA crash was observed as the SKB buffer address 
returned by the RX descriptor was corrupted and was 
pointing to an invalid address:
clean_rx: p: 14 c: 0 skb: 61763b10
WARN: src_info_type:0x0. Drop skb:61763b10

The reason for corruption seems to be the RX buffer allocation that was
happening twice before initiating the EDMA for the next transaction.
This change removes the additional allocation of the Rx buffer which is
not needed.

Change-Id: I4a5b404527469ff3b981749aa4e05080f55807cd
Signed-off-by: devi priya <quic_devipriy@quicinc.com>