tools: pack_v2: resolving wifi fw section name mis-match
This patch resolves the wififw section name mis-match for
the nor flash types.

In image info, wififw-SHA-HASH is used, whereas in the scripts
wifi_fw_ipq5332_qcn6432-SHA-HASH is used. It occur on nor-only build.
For nand/nor+nand, wifi usually in the ubi volume. Incase of
emmc, it is properly taken care.

So, handled all nor-only cases (including tiny-nor) as
similar as emmc case.

Change-Id: I3741a8255da487ad0546d9706933d19f1b1a3a93
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>